Convert ThemeContext to TypeScript
