+++ src/context/ThemeContext.tsx
@@ -1,0 +1,51 @@
+import { createContext, useContext, useEffect, useState, ReactNode } from 'react'
+
+export type Theme = 'light' | 'dark'
+
+export interface ThemeContextState {
+  theme: Theme
+  toggleTheme: () => void
+}
+
+const initialState: ThemeContextState = {
+  theme: 'light',
+  toggleTheme: () => {},
+}
+
+const ThemeContext = createContext<ThemeContextState>(initialState)
+
+export function ThemeProvider({ children }: { children: ReactNode }) {
+  const [theme, setTheme] = useState<Theme>('light')
+
+  const applyTheme = (nextTheme: Theme) => {
+    const root = document.documentElement
+    if (nextTheme === 'dark') {
+      root.classList.add('dark')
+    } else {
+      root.classList.remove('dark')
+    }
+  }
+
+  useEffect(() => {
+    const stored = localStorage.getItem('theme') as Theme | null
+    if (stored) {
+      setTheme(stored)
+      applyTheme(stored)
+    }
+  }, [])
+
+  const toggleTheme = () => {
+    const next: Theme = theme === 'light' ? 'dark' : 'light'
+    setTheme(next)
+    localStorage.setItem('theme', next)
+    applyTheme(next)
+  }
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  )
+}
+
+export const useTheme = () => useContext(ThemeContext)
